feat: MainPage 삭제하고 CreatePage, SearchPage로 분리
