+++ src/pages/home/subPages/create/hook/index.ts
@@ -1,0 +1,1 @@
+export { default as useCreateMemoManager } from './useCreateMemoManager';
