+++ src/pages/home/subPages/create/index.ts
@@ -1,0 +1,1 @@
+export { default as Create } from './Create';
