+++ src/pages/home/subPages/index.ts
@@ -1,4 +1,5 @@
-export * from './main';
+export * from './create';
+export * from './search';
 export * from './searchHistory';
 export * from './dashboard';
 export * from './uploadData';
