+++ src/pages/home/subPages/interfaces/Status/Status.ts
@@ -1,3 +1,1 @@
 export type Status = 'default' | 'loading' | 'success' | 'error';
-
-export type Mode = 'create' | 'search';
