+++ src/pages/home/subPages/interfaces/Status/index.ts
@@ -1,0 +1,1 @@
+export * from './Status';
