+++ src/pages/home/subPages/interfaces/index.ts
@@ -1,3 +1,4 @@
 export * from './MemoInterface';
 export * from './MemoSearch';
+export * from './Status';
 export * from './TagRelation';
