--- src/pages/home/subPages/main/components/ExamplesOrResultsAtSearchMode/index.ts
@@ -1,1 +1,0 @@
-export { default as ExamplesOrResultsAtSearchMode } from './ExamplesOrResultsAtSearchMode';
