--- src/pages/home/subPages/main/components/ModeToggle/index.ts
@@ -1,1 +1,0 @@
-export { default as ModeToggle } from './ModeToggle';
