--- src/pages/home/subPages/main/components/index.ts
@@ -1,4 +1,0 @@
-export * from './CreatedMemoList';
-export * from './ExamplesOrResultsAtSearchMode';
-export * from './MemoCreateTextArea';
-export * from './ModeToggle';
