--- src/pages/home/subPages/main/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from './type';
