--- src/pages/home/subPages/main/interfaces/type/index.ts
@@ -1,1 +1,0 @@
-export * from './type';
