+++ src/pages/home/subPages/search/ExamplesOrResults/ExampleContents/ExampleBox/ExampleBox.tsx
