+++ src/pages/home/subPages/search/ExamplesOrResults/ExampleContents/ExampleBox/index.ts
