+++ src/pages/home/subPages/search/ExamplesOrResults/ExampleContents/index.ts
