+++ src/pages/home/subPages/search/ExamplesOrResults/ResultContent/ResultContent.tsx
