+++ src/pages/home/subPages/search/ExamplesOrResults/ResultContent/index.ts
