+++ src/pages/home/subPages/search/ExamplesOrResults/index.ts
@@ -1,0 +1,1 @@
+export { default as ExamplesOrResults } from './ExamplesOrResults';
