+++ src/pages/home/subPages/search/hook/index.ts
@@ -1,2 +1,1 @@
-export { default as useCreateMemoManager } from './useCreateMemoManager';
 export { default as useSearchMemoManager } from './useSearchMemoManager';
