+++ src/pages/home/subPages/search/index.ts
@@ -1,0 +1,1 @@
+export { default as Search } from './Search';
